Move ConversationItem in separate file
