+++ src/features/conversations/overview/ConversationsOverview.tsx
@@ -4,63 +4,10 @@ import { Button } from '@hedvig-ui/Button/button'
 import { useConfirmDialog } from '@hedvig-ui/Modal/use-confirm-dialog'
 import { ConversationsRemaining } from 'features/conversations/overview/ConversationsRemaining'
 import { useMe } from 'features/user/hooks/use-me'
-import React, { useEffect, useRef } from 'react'
+import React, { useEffect } from 'react'
 import { useHistory } from 'react-router'
 import { QuestionGroup, UserSettingKey } from 'types/generated/graphql'
-
-const Item = styled(Flex)<{ selected: boolean }>`
-  background-color: ${({ theme, selected }) =>
-    selected ? theme.accent : theme.backgroundTransparent};
-
-  padding: 0 0.8em;
-  margin-top: 0.5em;
-  border-radius: 8px;
-  max-width: 100%;
-  width: 100%;
-  height: 2em;
-
-  cursor: pointer;
-
-  transition: all 300ms;
-
-  display: flex;
-  align-items: center;
-
-  & span {
-    font-size: 0.8em;
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    color: ${({ theme, selected }) =>
-      selected ? theme.accentContrast : theme.semiStrongForeground};
-  }
-
-  :first-of-type {
-    margin-top: 0.2em;
-  }
-
-  :hover {
-    background-color: ${({ theme }) => theme.accentLight};
-    color: ${({ theme }) => theme.accent};
-  }
-`
-
-interface ItemProps extends React.HTMLAttributes<HTMLDivElement> {
-  focus?: boolean
-  selected: boolean
-}
-
-const ConversationItem: React.FC<ItemProps> = ({ focus, ...props }) => {
-  const ref = useRef<HTMLDivElement>(null)
-
-  useEffect(() => {
-    if (focus && ref && ref.current) {
-      ref.current.focus()
-    }
-  }, [focus])
-
-  return <Item ref={ref} {...props} />
-}
+import { ConversationItem } from './ConversationItem'
 
 const ConversationWrapper = styled.div`
   margin-top: 1em;
@@ -130,16 +77,10 @@ export const ConversationsOverview: React.FC<{
         <ConversationWrapper>
           {filteredGroups.map((group, index) => (
             <ConversationItem
-              key={group.memberId}
-              tabIndex={0}
+              group={group}
+              currentMemberId={currentMemberId}
               focus={currentQuestionOrder === index}
-              onClick={() => history.push(`/conversations/${group.memberId}`)}
-              selected={group.memberId === currentMemberId}
-            >
-              <span>
-                {group.member?.firstName ?? ''} {group.member?.lastName ?? ' '}
-              </span>
-            </ConversationItem>
+            />
           ))}
         </ConversationWrapper>
       </Flex>
